requirements.txt
langchain
huggingface-hub
faiss-cpu
python-dotenv
CTransformers
sentence-transformers
